requirements.txt
Django==2.2.5
djangorestframework==3.10.3
mysqlclient==1.4.4
pkg-resources==0.0.0
pytz==2019.2
sqlparse==0.3.0
